# Performance: Fixes encoded strings performance