--- a/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/JsonRoundTripBenchmarks.cs
+++ b/Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Performance.Tests/Json/JsonRoundTripBenchmarks.cs
@@ -89,7 +89,6 @@
         {
             Text,
             Binary,
-            BinaryWithDictionaryEncoding,
             NewtonsoftText,
         }