# Pull Request Template

## Description
This PR optimizes encoded string performance by:
- improve the algorithm that identifies the charset to remove redundant loops in TryEncodeCompressedString
- keep track of the offsets at which string reference indexes are stored to avoid a full-pass over all values of the document (& cuts allocations by 50%)

Before:
| Method |     Serializer | EnableEncodedStrings |     Mean |   Error |  StdDev |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |--------------- |--------------------- |---------:|--------:|--------:|----------:|------:|------:|----------:|
|  Write |           Text |                False | 126.0 ms | 2.44 ms | 2.81 ms | 1000.0000 |     - |     - |  75.67 MB |
|  Write |           Text |                 True | 121.8 ms | 2.28 ms | 2.02 ms | 1000.0000 |     - |     - |  75.67 MB |
|  Write |    BinaryNaive |                False | 241.4 ms | 1.64 ms | 1.37 ms | 1000.0000 |     - |     - |  82.24 MB |
|  Write |    BinaryNaive |                 True | 515.9 ms | 5.00 ms | 4.43 ms | 3000.0000 |     - |     - | 183.48 MB |
|  Write | BinaryBlitting |                False | 150.8 ms | 3.01 ms | 7.21 ms | 1000.0000 |     - |     - |  71.71 MB |
|  Write | BinaryBlitting |                 True | 415.8 ms | 3.27 ms | 3.06 ms | 3000.0000 |     - |     - | 172.95 MB |

After:
| Method |     Serializer | EnableEncodedStrings |     Mean |   Error |  StdDev |   Median |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |--------------- |--------------------- |---------:|--------:|--------:|---------:|----------:|------:|------:|----------:|
|  Write |           Text |                False | 123.2 ms | 1.96 ms | 1.74 ms | 122.7 ms | 1000.0000 |     - |     - |  75.67 MB |
|  Write |           Text |                 True | 121.3 ms | 0.78 ms | 0.65 ms | 121.0 ms | 1000.0000 |     - |     - |  75.67 MB |
|  Write |    BinaryNaive |                False | 249.7 ms | 4.97 ms | 8.30 ms | 245.8 ms | 1000.0000 |     - |     - |  86.15 MB |
|  Write |    BinaryNaive |                 True | 278.5 ms | 4.75 ms | 4.45 ms | 278.4 ms | 1000.0000 |     - |     - |   99.1 MB |
|  Write | BinaryBlitting |                False | 147.2 ms | 2.89 ms | 4.67 ms | 146.7 ms | 1000.0000 |     - |     - |  73.54 MB |
|  Write | BinaryBlitting |                 True | 178.1 ms | 3.51 ms | 5.14 ms | 177.8 ms | 1000.0000 |     - |     - |  86.49 MB |

BinaryNaive exclusively uses public SDK primitives, whereas BinaryBlitting uses the typed JSON shortcuts. The EnableEncodedStrings has no impact on Text run.

### Benchmark Updates

#### Text Navigator to Binary Writer
##### Before 
|            Method |       payload | sourceFormat | destinationFormat |         Mean |      Error |     StdDev |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
|------------------ |-------------- |------------- |------------------ |-------------:|-----------:|-----------:|---------:|---------:|---------:|-----------:|
| NavigatorToWriter | NutritionData |                 Text |               Binary | 34,552.14 us | 619.581 us | 579.557 us | 1000.0000 | 333.3333 |        - | 5572.44 KB |

##### After
|            Method |       payload | sourceFormat | destinationFormat |         Mean |      Error |     StdDev |    Gen 0 |    Gen 1 |    Gen 2 |  Allocated |
|------------------ |-------------- |------------- |------------------ |-------------:|-----------:|-----------:|---------:|---------:|---------:|-----------:|
| NavigatorToWriter | NutritionData |         Text |            Binary | 24,779.95 us | 515.386 us | 755.445 us | 875.0000 | 500.0000 | 125.0000 | 5053.33 KB |

30% reduction in time and 10% reduction in memory

#### Binary String Write MicroBenchmark

##### Before 
|                     Method | namedWriteDelegate | benchmarkSerializationFormat |      Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|--------------------------- |------------------- |----------------------------- |----------:|----------:|----------:|-----------:|----------:|----------:|----------:|
| ExecuteWriteMicroBenchmark |             string |                       Binary | 1,023.25 ms | 9.114 ms | 8.525 ms | 35000.0000 | 2000.0000 | 2000.0000 | 145.39 MB |

##### After
|                     Method | namedWriteDelegate | benchmarkSerializationFormat |      Mean |     Error |    StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|--------------------------- |------------------- |----------------------------- |----------:|----------:|----------:|-----------:|----------:|----------:|----------:|
| ExecuteWriteMicroBenchmark |             string |                       Binary | 452.70 ms |  8.968 ms | 15.229 ms | 36000.0000 | 2000.0000 | 2000.0000 | 161.39 MB |

half the time

## Type of change


## Closing issues

To automatically close an issue: closes #IssueNumber