@bchong95 2020-12-17
> This is not Internal though, right? If it would benefit end users, so can we change the title?

No user uses binary writer, so it's still internal.